+++ lib/dormouse/views/form.rb
@@ -4,6 +4,7 @@ class Dormouse::Views::Form < Dormouse::Views::Base
     manifest = self.manifest
     
     widgets  = self.widgets
+    tabs     = self.tabs
     
     object ||= begin
       if id = @controller.params[:id]
@@ -16,6 +17,7 @@ class Dormouse::Views::Form < Dormouse::Views::Base
     controller_eval do
       @object   = object
       @widgets  = widgets
+      @tabs     = tabs
       render :template => "#{manifest.style}/views/form", :layout => "#{manifest.style}/layouts/dormouse"
     end
   end
@@ -25,6 +27,19 @@ class Dormouse::Views::Form < Dormouse::Views::Base
       manifest.inject([]) do |memo, property|
         widget_klass = Dormouse::Widgets[property.type]
         memo << widget_klass.new(manifest, property) if widget_klass
+        memo
+      end
+    end
+  end
+  
+  def tabs
+    @tabs ||= begin
+      manifest.inject([]) do |memo, property|
+        
+        if property.type == :has_many and !property.options[:inline]
+          memo << Dormouse::Tab.new(manifest, property)
+        end
+        
         memo
       end
     end